optimize for main and change name to scraper.py
